+++ KFXLog/Classes/View Controllers/KFXLogFileDetailVC.m
@@ -36,9 +36,19 @@
 
 @interface KFXLogFileDetailVC () <MFMailComposeViewControllerDelegate, UISearchBarDelegate>
 
-@property (strong, nonatomic)  UITextView *textView;
+// Text
 @property (strong,nonatomic) NSString *filePath;
+@property (strong,nonatomic) NSMutableArray <NSValue*>*matchingRanges;
+@property (nonatomic) NSUInteger selectedMatch;
+// UI
 @property (strong,nonatomic) UISearchBar *searchBar;
+@property (strong,nonatomic) UIView *resultBar;
+@property (strong,nonatomic) UITextView *textView;
+@property (strong,nonatomic) UIButton *nextResultButton;
+@property (strong,nonatomic) UIButton *previousResultButton;
+@property (strong,nonatomic) UILabel *resultCountLabel;
+@property (strong,nonatomic) NSLayoutConstraint *searchBarOriginY;
+
 @end
 
 @implementation KFXLogFileDetailVC
@@ -66,25 +76,89 @@
 - (void)viewDidLoad {
     [super viewDidLoad];
     self.title = @"";
+    [KFXLog logInfoWithSender:self format:@"Log File Viewer opened for log file at path: %@",self.filePath];
+    [self addSubviews];
 }
 
 -(void)viewWillAppear:(BOOL)animated{
     [super viewWillAppear:animated];
     
-    // ## Add Text View ##
-    CGRect textViewFrame = self.view.bounds;
-    self.textView = [[UITextView alloc]initWithFrame:textViewFrame];
-    self.textView.autoresizingMask = UIViewAutoresizingFlexibleWidth | UIViewAutoresizingFlexibleHeight;
-    self.textView.textContainerInset = UIEdgeInsetsMake(50.0, 0.0, 0.0, 0.0);
-    [self.view addSubview:self.textView];
+
+    // ## Display file log text ##
+    if (self.filePath != nil) {
+        [self refreshTextViewText];
+    }
+
+    [self addConstraintsForSubviews];
+}
+
+-(void)viewDidAppear:(BOOL)animated{
+    [super viewDidAppear:animated];
+}
+
+- (void)didReceiveMemoryWarning {
+    [super didReceiveMemoryWarning];
+}
+
+-(void)addSubviews{
+    
     
     // ## Add Search Bar ##
     self.searchBar = [[UISearchBar alloc]init];
     self.searchBar.delegate = self;
     self.searchBar.returnKeyType = UIReturnKeyDone;
+    self.searchBar.showsCancelButton = YES;
     self.searchBar.translatesAutoresizingMaskIntoConstraints = NO;
     [self.view addSubview:self.searchBar];
+    
+    // ## Add Result Bar ##
+    self.resultBar = [[UIView alloc]init];
+    self.resultBar.backgroundColor = [UIColor lightGrayColor];
+    self.resultBar.translatesAutoresizingMaskIntoConstraints = NO;
+    [self.view addSubview:self.resultBar];
+    // Buttons
+    UIButton *nextResultButton = [UIButton buttonWithType:UIButtonTypeRoundedRect];
+    [nextResultButton setTitle:@">" forState:UIControlStateNormal];
+    [nextResultButton setTitleColor:[UIColor redColor] forState:UIControlStateNormal];
+    nextResultButton.titleLabel.font = [UIFont systemFontOfSize:30.0 weight:UIFontWeightBold];
+    nextResultButton.enabled = NO;
+    nextResultButton.translatesAutoresizingMaskIntoConstraints = NO;
+    [self.resultBar addSubview:nextResultButton];
+    self.nextResultButton = nextResultButton;
+    [self.nextResultButton addTarget:self
+                              action:@selector(nextResultButtonTapped:)
+                    forControlEvents:UIControlEventTouchUpInside];
+    UIButton *previousResultButton = [UIButton buttonWithType:UIButtonTypeRoundedRect];
+    [previousResultButton setTitle:@"<" forState:UIControlStateNormal];
+    [previousResultButton setTitleColor:[UIColor redColor] forState:UIControlStateNormal];
+    previousResultButton.titleLabel.font = [UIFont systemFontOfSize:30.0 weight:UIFontWeightBold];
+    previousResultButton.enabled = NO;
+    previousResultButton.translatesAutoresizingMaskIntoConstraints = NO;
+    [self.resultBar addSubview:previousResultButton];
+    self.previousResultButton = previousResultButton;
+    [self.previousResultButton addTarget:self
+                              action:@selector(previousResultButtonTapped:)
+                    forControlEvents:UIControlEventTouchUpInside];
 
+    // Label
+    UILabel *resultCountLabel = [[UILabel alloc]init];
+    resultCountLabel.font = [UIFont systemFontOfSize:15.0 weight:UIFontWeightMedium];
+    resultCountLabel.textColor = [UIColor blackColor];
+    resultCountLabel.text = @"Use the search bar to search the logs";
+    resultCountLabel.translatesAutoresizingMaskIntoConstraints = NO;
+    [self.resultBar addSubview:resultCountLabel];
+    self.resultCountLabel = resultCountLabel;
+
+    // ## Add Text View ##
+    self.textView = [[UITextView alloc]init];
+    //    self.textView.textContainerInset = UIEdgeInsetsMake(50.0, 0.0, 0.0, 0.0);
+    self.textView.backgroundColor = [UIColor whiteColor];
+    self.textView.editable = NO;
+    self.textView.selectable = YES;
+    self.textView.translatesAutoresizingMaskIntoConstraints = NO;
+    [self.view addSubview:self.textView];
+    
+    // ## Add Navigation Bar buttons ##
     if (self.navigationController.navigationBar != nil) {
         // ## Add Bar Buttons ##
         UIBarButtonItem *emailButton = [[UIBarButtonItem alloc]initWithTitle:@"Email" style:UIBarButtonItemStylePlain target:self action:@selector(emailButtonTapped:)];
@@ -95,24 +169,18 @@
         self.navigationItem.rightBarButtonItems = @[emailButton,refreshButton,jumpToEndButton,jumpToStartButton];
     }
 
-    // ## Display file log text ##
-    if (self.filePath != nil) {
-        [self refreshTextViewText];
-    }
-
-    // The search bar needs to have constraints set
-    [self.view setNeedsUpdateConstraints];
 }
 
-- (void)didReceiveMemoryWarning {
-    [super didReceiveMemoryWarning];
+-(void)addConstraintsForSubviews{
     
-}
-
--(void)updateViewConstraints{
-    
-    [self.view removeConstraints:self.view.constraints];
-    NSDictionary *viewsDict = @{@"searchBar":self.searchBar};
+    NSDictionary *viewsDict = @{
+                                @"searchBar":self.searchBar,
+                                @"resultBar":self.resultBar,
+                                @"textView":self.textView,
+                                @"resultLabel":self.resultCountLabel,
+                                @"prevButton":self.previousResultButton,
+                                @"nextButton":self.nextResultButton
+                                };
     CGFloat searchBarOriginY = 0.0;
     if (self.navigationController.navigationBar != nil) {
         searchBarOriginY += self.navigationController.navigationBar.bounds.size.height;
@@ -120,27 +188,81 @@
     if (![UIApplication sharedApplication].isStatusBarHidden) {
         searchBarOriginY += 20.0;
     }
+    NSDictionary *metrics = @{
+                              @"searchBarY":@(searchBarOriginY),
+                              @"labelHeight":@(30)
+                              };
     
-    NSDictionary *metrics = @{@"searchBarY":@(searchBarOriginY)};
+    // ## Horizontal ##
     [self.view addConstraints:[NSLayoutConstraint constraintsWithVisualFormat:@"H:|[searchBar]|"
                                                                       options:kNilOptions
                                                                       metrics:nil
                                                                         views:viewsDict]];
-    [self.view addConstraints:[NSLayoutConstraint constraintsWithVisualFormat:@"V:|-searchBarY-[searchBar(50)]"
+    [self.view addConstraints:[NSLayoutConstraint constraintsWithVisualFormat:@"H:|[resultBar]|"
+                                                                      options:kNilOptions
+                                                                      metrics:nil
+                                                                        views:viewsDict]];
+    [self.view addConstraints:[NSLayoutConstraint constraintsWithVisualFormat:@"H:|[textView]|"
+                                                                      options:kNilOptions
+                                                                      metrics:nil
+                                                                        views:viewsDict]];
+    [self.resultBar addConstraints:[NSLayoutConstraint constraintsWithVisualFormat:@"H:|-20-[resultLabel]-[prevButton][nextButton]-|"
+                                                                           options:kNilOptions
+                                                                           metrics:metrics
+                                                                             views:viewsDict]];
+    
+    // ## Vertical ##
+    self.searchBarOriginY = [NSLayoutConstraint constraintWithItem:self.searchBar
+                                                         attribute:NSLayoutAttributeTop
+                                                         relatedBy:NSLayoutRelationEqual
+                                                            toItem:self.view
+                                                         attribute:NSLayoutAttributeTop
+                                                        multiplier:1.0
+                                                          constant:searchBarOriginY];
+    [self.view addConstraint:self.searchBarOriginY];
+    [self.view addConstraints:[NSLayoutConstraint constraintsWithVisualFormat:@"V:[searchBar(50)][resultBar(45)][textView]|"
                                                                       options:kNilOptions
                                                                       metrics:metrics
                                                                         views:viewsDict]];
-    
+    [self.resultBar addConstraints:[NSLayoutConstraint constraintsWithVisualFormat:@"V:|[resultLabel]|"
+                                                                      options:kNilOptions
+                                                                      metrics:metrics
+                                                                        views:viewsDict]];
+    [self.resultBar addConstraints:[NSLayoutConstraint constraintsWithVisualFormat:@"V:|[prevButton]|"
+                                                                      options:kNilOptions
+                                                                      metrics:nil
+                                                                        views:viewsDict]];
+    [self.resultBar addConstraints:[NSLayoutConstraint constraintsWithVisualFormat:@"V:|[nextButton]|"
+                                                                      options:kNilOptions
+                                                                      metrics:nil
+                                                                        views:viewsDict]];
+}
 
+-(void)updateViewConstraints{
+
+    /*
+     Update the constraint for the search bar origin Y based on if there is a status bar and navigation bar visible - keep search bar flush to the top subview
+    */
+    CGFloat searchBarOriginY = 0.0;
+    if (self.navigationController.navigationBar != nil) {
+        searchBarOriginY += self.navigationController.navigationBar.bounds.size.height;
+    }
+    if (![UIApplication sharedApplication].isStatusBarHidden) {
+        searchBarOriginY += 20.0;
+    }
+    self.searchBarOriginY.constant = searchBarOriginY;
+    
     [super updateViewConstraints];
 }
+
+
 
 //--------------------------------------------------------
 #pragma mark - Actions
 //--------------------------------------------------------
--(IBAction)emailButtonTapped:(id)sender{
+-(void)emailButtonTapped:(id)sender{
     
-    
+    [KFXLog logUIEventWithSender:self format:@"-emailButtonTapped:"];
     if (![MFMailComposeViewController canSendMail]) {
         [KFXLog logFail:@"Cannot send mail from this device"];
         return;
@@ -150,7 +272,7 @@
     NSString *subject = [self.filePath lastPathComponent];
     
     MFMailComposeViewController *mailVC = [[MFMailComposeViewController alloc]init];
-    [mailVC setMailComposeDelegate:self]; // Note the delegate is called MailComposeDelegate
+    [mailVC setMailComposeDelegate:self];
     [mailVC setSubject:subject];
     NSData *data = [[NSFileManager defaultManager]contentsAtPath:self.filePath];
     if (data == nil) {
@@ -164,20 +286,92 @@
     [self presentViewController:mailVC animated:YES completion:NULL];
 }
 
--(IBAction)refreshButtonTapped:(id)sender{
-    
+-(void)refreshButtonTapped:(id)sender{
+    [KFXLog logUIEventWithSender:self format:@"-refreshButtonTapped: *-*-*-*-*-*-*-*-*-*:"];
+
     [self refreshTextViewText];
     
 }
 
--(IBAction)jumpToTopButtonTapped:(id)sender{
+-(void)jumpToTopButtonTapped:(id)sender{
     [self.textView scrollRangeToVisible:NSMakeRange(0, 0)];
 }
 
--(IBAction)jumpToBottomButtonTapped:(id)sender{
+-(void)jumpToBottomButtonTapped:(id)sender{
     [self.textView scrollRangeToVisible:NSMakeRange([self.textView.text length], 0)];
     
 }
+
+-(void)previousResultButtonTapped:(UIButton*)sender{
+
+    if (self.selectedMatch >= 1) {
+        
+        NSMutableAttributedString *mutAttString = [self.textView.attributedText mutableCopy];
+        
+        // Reset the previously selected match
+        NSValue *oldValue = self.matchingRanges[self.selectedMatch];
+        NSRange oldRange = [oldValue rangeValue];
+        [self highlightRange:oldRange ofAttributedString:mutAttString withColour:[UIColor greenColor]];
+
+        // update
+        self.selectedMatch--;
+        
+        // Highlight new selected match
+        NSValue *newValue = self.matchingRanges[self.selectedMatch];
+        NSRange newRange = [newValue rangeValue];
+        [self highlightRange:newRange ofAttributedString:mutAttString withColour:[UIColor purpleColor]];
+        
+        // update label
+        [self updateResultCountLabel];
+        
+        // Update text view
+        self.textView.attributedText = [mutAttString copy];
+        [self scrollTextViewToTextInRange:newRange];
+    }
+}
+
+-(void)nextResultButtonTapped:(UIButton*)sender{
+    
+    if (self.selectedMatch+1 < self.matchingRanges.count) {
+        
+        NSMutableAttributedString *mutAttString = [self.textView.attributedText mutableCopy];
+        
+        // Reset the previously selected match
+        NSValue *oldValue = self.matchingRanges[self.selectedMatch];
+        NSRange oldRange = [oldValue rangeValue];
+        [self highlightRange:oldRange ofAttributedString:mutAttString withColour:[UIColor greenColor]];
+        
+        // update
+        self.selectedMatch++;
+        
+        // Highlight new selected match
+        NSValue *newValue = self.matchingRanges[self.selectedMatch];
+        NSRange newRange = [newValue rangeValue];
+        [self highlightRange:newRange ofAttributedString:mutAttString withColour:[UIColor purpleColor]];
+        
+        // update label
+        [self updateResultCountLabel];
+        
+        // Update text view
+        self.textView.attributedText = [mutAttString copy];
+        [self scrollTextViewToTextInRange:newRange];
+    }
+}
+
+-(void)scrollTextViewToTextInRange:(NSRange)range{
+
+    [self.textView setSelectedRange:range];
+    
+    CGRect textRect = [self.textView firstRectForRange:self.textView.selectedTextRange];
+    [KFXLog logInfo:@"TextRect: %@",NSStringFromCGRect(textRect)];
+    if (!CGRectContainsRect(self.textView.bounds, textRect)) {
+        [self.textView scrollRectToVisible:textRect
+                                  animated:YES];
+
+    }
+    
+    [self.textView setSelectedRange:NSMakeRange(0, 0)];
+    }
 
 //======================================================
 #pragma mark - ** Protocol Methods **
@@ -190,16 +384,16 @@
     [self dismissViewControllerAnimated:YES completion:nil];
     switch (result) {
         case MFMailComposeResultSent:
-            [KFXLog logInfo:@"Sent Log File by Email"];
+            [KFXLog logInfoWithSender:self format:@"Email log file: Sent"];
             break;
         case MFMailComposeResultSaved:
-            [KFXLog logInfo:@"Saved email with log file"];
+            [KFXLog logInfoWithSender:self format:@"Email log file: Saved"];
             break;
         case MFMailComposeResultFailed:
-            [KFXLog logInfo:@"Log file sending by email failed"];
+            [KFXLog logInfoWithSender:self format:@"Email log file: Failed"];
             break;
         case MFMailComposeResultCancelled:
-            [KFXLog logInfo:@"Log file email sending cancelled"];
+            [KFXLog logInfoWithSender:self format:@"Email log file: Cancelled"];
             break;
             
         default:
@@ -218,6 +412,7 @@
 -(void)searchBar:(UISearchBar *)searchBar textDidChange:(NSString *)searchText{
     
     [self highlightInstancesOfString:searchText];
+    [self updateResultCountLabel];
 }
 
 -(void)searchBarSearchButtonClicked:(UISearchBar *)searchBar{
@@ -226,6 +421,13 @@
 
 -(void)searchBarTextDidEndEditing:(UISearchBar *)searchBar{
     
+}
+
+-(void)searchBarCancelButtonClicked:(UISearchBar *)searchBar{
+    searchBar.text = nil;
+    [self highlightInstancesOfString:@""];
+    [self updateResultCountLabel];
+    [searchBar resignFirstResponder];
 }
 
 
@@ -273,6 +475,10 @@
     [mutAttString removeAttribute:NSBackgroundColorAttributeName range:NSMakeRange(0, mutAttString.length)];
     self.textView.attributedText = [mutAttString copy];
     
+    // Clear previous matches
+    [self.matchingRanges removeAllObjects];
+    self.selectedMatch = 0;
+    
     NSString *haystack = self.textView.attributedText.string;
 
     // From: http://stackoverflow.com/questions/7033574/find-all-locations-of-substring-in-nsstring-not-just-first
@@ -285,26 +491,56 @@
         if (foundRange.location != NSNotFound) {
             // found an occurrence of the substring! do stuff here
             searchRange.location = foundRange.location+foundRange.length;
-            [mutAttString beginEditing];
-            [mutAttString addAttribute:NSBackgroundColorAttributeName
-                              value:[UIColor greenColor]
-                              range:foundRange];
-            [mutAttString endEditing];
+            [self highlightRange:foundRange ofAttributedString:mutAttString withColour:[UIColor greenColor]];
+            [self.matchingRanges addObject:[NSValue valueWithRange:foundRange]];
         } else {
             // no more substring to find
             self.textView.attributedText = [mutAttString copy];
             break;
         }
     }
-
-    
+    if (self.matchingRanges.count >= 1) {
+        self.previousResultButton.enabled = YES;
+        self.nextResultButton.enabled = YES;
+    }else{
+        self.previousResultButton.enabled = NO;
+        self.nextResultButton.enabled = NO;
+    }
 }
 
+-(void)highlightRange:(NSRange)range ofAttributedString:(NSMutableAttributedString*)mutAttString withColour:(UIColor*)colour{
+    
+    [mutAttString beginEditing];
+    [mutAttString addAttribute:NSBackgroundColorAttributeName
+                         value:colour
+                         range:range];
+    [mutAttString endEditing];
+
+}
+
+
+-(void)updateResultCountLabel{
+    
+    NSUInteger selectedMatch = self.selectedMatch;
+    NSUInteger matchesCount = self.matchingRanges.count;
+    if (matchesCount == 0) {
+        selectedMatch = 0;
+    }else{
+        selectedMatch++;
+    }
+    
+    self.resultCountLabel.text = [NSString stringWithFormat:@"%lu of %lu matches",(unsigned long)selectedMatch,(unsigned long)matchesCount];
+}
 
 //--------------------------------------------------------
 #pragma mark - Lazy Load
 //--------------------------------------------------------
-
+-(NSMutableArray <NSValue*>*)matchingRanges{
+    if (!_matchingRanges) {
+        _matchingRanges = [NSMutableArray arrayWithCapacity:100];
+    }
+    return _matchingRanges;
+}